(PR changes): Dict and Model factories one file
Refactor tests affected by this.
